修正I2c BUSY问题，解决BMP180I2C写数据错误
修正I2c BUSY问题，解决BMP180I2C写数据错误

+++ hello_lcd/Drivers/BSP/STM32ZET6_MILESTONE/stm32zet6_milestone_lcd.h
@@ -179,7 +179,7 @@ void     BSP_LCD_FillEllipse(int Xpos, int Ypos, int XRadius, int YRadius);
 void     BSP_LCD_DisplayOff(void);
 void     BSP_LCD_DisplayOn(void);
 void     BSP_LCD_DisplayStringAtXposLine(uint16_t x_pos,uint16_t Line, uint8_t *pText);
-void     BSP_LCD_DisplayUint16DecAtXposLine(uint16_t x_pos ,uint16_t Line,uint16_t digit,uint8_t cnt);
+void     BSP_LCD_DisplayUint16DecAtXposLine(uint16_t x_pos ,uint16_t Line,uint32_t digit,uint8_t cnt);
 void     BSP_LCD_DisplayUint16HexAtXposLine(uint16_t x_pos ,uint16_t Line,uint16_t digital);
 void     BSP_LCD_DisplayUint32HexAtXposLine(uint16_t x_pos ,uint16_t Line,uint32_t digital);
 void     BSP_LCD_DisplayfloatAtXposLine(uint16_t x_pos ,uint16_t Line,float fdigit,uint8_t cnt);
